Rename top directopry to index
